add explorer support to vscode
