+++ src/extension/tab/hook.ts
@@ -1,25 +1,12 @@
-import type { ApolloClient, ApolloError } from "@apollo/client";
-
-// Note that we are intentionally not using Apollo Client's gql and
-// Observable exports, as we don't want Apollo Client and its dependencies
-// to be loaded into each browser tab, when this hook triggered.
-import Observable from "zen-observable";
-import type { DefinitionNode } from "graphql/language";
-
-type Writable<T> = { -readonly [P in keyof T]: T[P] };
+import type { ApolloClient } from "@apollo/client";
 
 // All manifests should contain the same version number so it shouldn't matter
 // which one we import from.
 import * as manifest from "../chrome/manifest.json";
 const { version: devtoolsVersion } = manifest;
 import type { MutationDetails, QueryDetails } from "./helpers";
-import {
-  getQueries,
-  getQueriesLegacy,
-  getMutations,
-  getMainDefinition,
-} from "./helpers";
-import type { ApolloClientInfo, QueryResult, SafeAny } from "../../types";
+import { getQueries, getQueriesLegacy, getMutations } from "./helpers";
+import type { ApolloClientInfo, SafeAny } from "../../types";
 import { getPrivateAccess } from "../../privateAccess";
 import type { JSONObject } from "../../application/types/json";
 import { createWindowActor } from "../actor";
@@ -27,6 +14,7 @@ import { createWindowMessageAdapter } from "../messageAdapters";
 import { createRpcClient, createRpcHandler } from "../rpc";
 import { loadErrorCodes } from "./loadErrorCodes";
 import { createId } from "../../utils/createId";
+import { handleExplorerRequests } from "./handleExplorerRequests";
 
 declare global {
   type TCache = any;
@@ -150,102 +138,7 @@ function getClientById(clientId: string) {
   return client;
 }
 
-tab.on("explorerRequest", (message) => {
-  const {
-    clientId,
-    operation: queryAst,
-    operationName,
-    fetchPolicy,
-    variables,
-  } = message.payload;
-
-  const client = getClientById(clientId);
-
-  if (!client) {
-    throw new Error("Could not find selected client");
-  }
-
-  const clonedQueryAst = structuredClone(queryAst) as Writable<typeof queryAst>;
-
-  const filteredDefinitions = clonedQueryAst.definitions.reduce(
-    (acumm: DefinitionNode[], curr) => {
-      if (
-        (curr.kind === "OperationDefinition" &&
-          curr.name?.value === operationName) ||
-        curr.kind !== "OperationDefinition"
-      ) {
-        acumm.push(curr);
-      }
-
-      return acumm;
-    },
-    []
-  );
-
-  clonedQueryAst.definitions = filteredDefinitions;
-
-  const definition = getMainDefinition(clonedQueryAst);
-
-  const operation = (() => {
-    if (
-      definition.kind === "OperationDefinition" &&
-      definition.operation === "mutation"
-    ) {
-      return new Observable<QueryResult>((observer) => {
-        client
-          .mutate({ mutation: clonedQueryAst, variables })
-          .then((result) => {
-            observer.next(result as QueryResult);
-          });
-      });
-    } else {
-      return client.watchQuery({
-        query: clonedQueryAst,
-        variables,
-        fetchPolicy,
-      });
-    }
-  })();
-
-  const operationObservable = operation?.subscribe(
-    (response: QueryResult) => {
-      tab.send({
-        type: "explorerResponse",
-        payload: { operationName, response },
-      });
-    },
-    (error: ApolloError) => {
-      tab.send({
-        type: "explorerResponse",
-        payload: {
-          operationName,
-          response: {
-            errors: error.graphQLErrors.length
-              ? error.graphQLErrors
-              : error.networkError && "result" in error.networkError
-                ? typeof error.networkError?.result === "string"
-                  ? error.networkError?.result
-                  : error.networkError?.result.errors ?? []
-                : [],
-            error: error,
-            data: null,
-            loading: false,
-            networkStatus: 8, // NetworkStatus.error - we want to prevent importing the enum here
-          },
-        },
-      });
-    }
-  );
-
-  if (
-    definition.kind === "OperationDefinition" &&
-    definition.operation === "subscription"
-  ) {
-    tab.on("explorerSubscriptionTermination", () => {
-      operationObservable?.unsubscribe();
-    });
-  }
-});
+handleExplorerRequests(tab, getClientById);
 
 function watchForClientTermination(client: ApolloClient<any>) {
   const originalStop = client.stop;
